Make `nonnative` a submodule of `fields`
