+++ src/fields/mod.rs
@@ -20,6 +20,10 @@ pub mod quadratic_extension;
 /// That is, it implements the R1CS equivalent of `ark_ff::Fp*`.
 pub mod fp;
 
+/// This module contains a generic implementation of "nonnative" prime field variables.
+/// It emulates `Fp` arithmetic over `Fq` arithmetic, where `p != q`.
+pub mod nonnative;
+
 /// This module contains a generic implementation of the degree-12 tower
 /// extension field. That is, it implements the R1CS equivalent of
 /// `ark_ff::Fp12`
