--- src/fields/nonnative/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,26 +1,0 @@
-<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
-v                               ✰  Thanks for creating a PR! ✰
-v    Before hitting that submit button please review the checkboxes.
-v    If a checkbox is n/a - please still include it but + a little note why
-☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->
-
-## Description
-
-<!-- Add a description of the changes that this PR introduces and the files that
-are the most critical to review.
--->
-
-closes: #XXXX
-
----
-
-Before we can merge this PR, please make sure that all the following items have been
-checked off. If any of the checklist items are not applicable, please leave them but
-write a little note why.
-
-- [ ] Targeted PR against correct branch (master)
-- [ ] Linked to Github issue with discussion and accepted design OR have an explanation in the PR that describes this work.
-- [ ] Wrote unit tests
-- [ ] Updated relevant documentation in the code
-- [ ] Added a relevant changelog entry to the `Pending` section in `CHANGELOG.md`
-- [ ] Re-reviewed `Files changed` in the Github PR explorer
