--- src/fields/nonnative/.github/dependabot.yml
@@ -1,11 +1,0 @@
-version: 2
-updates:
-- package-ecosystem: cargo
-  directory: "/"
-  schedule:
-    interval: daily
-  open-pull-requests-limit: 10
-  ignore:
-  - dependency-name: rand
-    versions:
-    - 0.8.0
