--- src/fields/nonnative/.gitignore
@@ -1,12 +1,0 @@
-target
-Cargo.lock
-.DS_Store
-.idea
-*.iml
-*.ipynb_checkpoints
-*.pyc
-*.sage.py
-params
-*.swp
-*.swo
-
