--- src/fields/nonnative/CHANGELOG.md
@@ -1,27 +1,0 @@
-# CHANGELOG
-
-## Pending
-
-### Breaking changes
-
-### Features
-
-- [\#45](https://github.com/arkworks-rs/nonnative/pull/45) Add `new_witness_with_le_bits` which returns the bits used during variable allocation.
-
-### Improvements
-
-### Bug fixes
-
-## v0.3.0
-
-### Breaking changes
-
-- [\#43](https://github.com/arkworks-rs/nonnative/pull/43) Add padding to allocated nonnative element's `to_bytes`.
-
-### Features
-
-### Improvements
-
-### Bug fixes
-
-## v0.2.0 (Initial release of ark-nonnative)
