--- src/fields/nonnative/Cargo.toml
@@ -1,71 +1,0 @@
-[package]
-name = "ark-nonnative-field"
-version = "0.3.1"
-authors = [
-    "Weikeng Chen",
-    "Alessandro Chiesa",
-    "Emma Dauterman",
-    "Nicholas Ward"
-]
-description = "Constraints for nonnative field gadgets"
-homepage = "https://arkworks.rs"
-repository = "https://github.com/arkworks-rs/nonnative"
-documentation = "https://docs.rs/ark-nonnative-field/"
-keywords = ["r1cs", "nonnative", "finite-fields"]
-categories = ["cryptography"]
-include = ["Cargo.toml", "src", "tests", "README.md"]
-license = "MIT/Apache-2.0"
-edition = "2018"
-
-[profile.release]
-debug = true
-panic = 'abort'
-
-[profile.test]
-opt-level = 3
-lto = "thin"
-incremental = true
-
-[profile.bench]
-opt-level = 3
-lto = "fat"
-incremental = true
-debug = false
-rpath = false
-debug-assertions = false
-
-################################# Dependencies ################################
-
-[dependencies]
-derivative = { version = "2", features = [ "use_core" ] }
-
-tracing = { version = "0.1", default-features = false, features = [ "attributes" ] }
-
-ark-ff = { version = "^0.3.0", default-features = false }
-ark-ec = { version = "^0.3.0", default-features = false }
-ark-std = { version = "^0.3.0", default-features = false }
-ark-relations = { version = "^0.3.0", default-features = false }
-ark-r1cs-std = { version = "^0.3.0", default-features = false  }
-
-num-traits = { version = "0.2", default-features = false }
-num-bigint = { version = "0.4.0", default-features = false }
-num-integer = { version = "0.1.44", default-features = false }
-
-[dev-dependencies]
-paste = "1.0"
-ark-bls12-377 = { version = "^0.3.0", features = ["curve"], default-features = false  }
-ark-bls12-381 = { version = "^0.3.0", features = ["curve"], default-features = false  }
-ark-mnt4-298 = { version = "^0.3.0", features = ["curve"], default-features = false  }
-ark-mnt4-753 = { version = "^0.3.0", features = ["curve"], default-features = false  }
-ark-mnt6-298 = { version = "^0.3.0", default-features = false  }
-ark-mnt6-753 = { version = "^0.3.0", default-features = false  }
-ark-pallas = { version = "^0.3.0", features = ["curve"],  default-features = false  }
-
-[features]
-default = []
-std = [ "ark-std/std", "ark-ff/std", "ark-ec/std", "ark-relations/std", "ark-r1cs-std/std", "num-traits/std", "num-bigint/std" ]
-
-[[bench]]
-name = "nonnative-bench"
-path = "benches/bench.rs"
-harness = false
