+++ src/fields/nonnative/allocated_nonnative_field_var.rs
@@ -1,10 +1,10 @@
-use crate::params::{get_params, OptimizationType};
-use crate::reduce::{bigint_to_basefield, limbs_to_bigint, Reducer};
-use crate::AllocatedNonNativeFieldMulResultVar;
+use super::params::{get_params, OptimizationType};
+use super::reduce::{bigint_to_basefield, limbs_to_bigint, Reducer};
+use super::AllocatedNonNativeFieldMulResultVar;
 use ark_ff::{BigInteger, FpParameters, PrimeField};
-use ark_r1cs_std::fields::fp::FpVar;
-use ark_r1cs_std::prelude::*;
-use ark_r1cs_std::ToConstraintFieldGadget;
+use crate::fields::fp::FpVar;
+use crate::prelude::*;
+use crate::ToConstraintFieldGadget;
 use ark_relations::r1cs::{OptimizationGoal, Result as R1CSResult};
 use ark_relations::{
     ns,
