+++ src/fields/nonnative/nonnative_field_mul_result_var.rs
@@ -1,4 +1,4 @@
-use crate::{AllocatedNonNativeFieldMulResultVar, NonNativeFieldVar};
+use super::{AllocatedNonNativeFieldMulResultVar, NonNativeFieldVar};
 use ark_ff::PrimeField;
 use ark_relations::r1cs::Result as R1CSResult;
