+++ src/fields/nonnative/nonnative_field_var.rs
@@ -1,12 +1,12 @@
-use crate::params::OptimizationType;
-use crate::{AllocatedNonNativeFieldVar, NonNativeFieldMulResultVar};
+use super::params::OptimizationType;
+use super::{AllocatedNonNativeFieldVar, NonNativeFieldMulResultVar};
 use ark_ff::PrimeField;
 use ark_ff::{to_bytes, FpParameters};
-use ark_r1cs_std::boolean::Boolean;
-use ark_r1cs_std::fields::fp::FpVar;
-use ark_r1cs_std::fields::FieldVar;
-use ark_r1cs_std::prelude::*;
-use ark_r1cs_std::{R1CSVar, ToConstraintFieldGadget};
+use crate::boolean::Boolean;
+use crate::fields::fp::FpVar;
+use crate::fields::FieldVar;
+use crate::prelude::*;
+use crate::{R1CSVar, ToConstraintFieldGadget};
 use ark_relations::r1cs::Result as R1CSResult;
 use ark_relations::r1cs::{ConstraintSystemRef, Namespace, SynthesisError};
 use ark_std::hash::{Hash, Hasher};
